backend-app-api: restructure implementation layout
