+++ .changeset/twenty-moons-sneeze.md
@@ -1,0 +1,5 @@
+---
+'@backstage/backend-app-api': patch
+---
+
+Internal folder structure refactor.
