+++ packages/backend-app-api/src/services/implementations/config/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Backstage Authors
+ * Copyright 2023 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,17 +14,4 @@
  * limitations under the License.
  */
 
-export * from './cache';
-export * from './config';
-export * from './database';
-export * from './discovery';
-export * from './httpRouter';
-export * from './lifecycle';
-export * from './logger';
-export * from './permissions';
-export * from './rootHttpRouter';
-export * from './rootLifecycle';
-export * from './rootLogger';
-export * from './scheduler';
-export * from './tokenManager';
-export * from './urlReader';
+export { configFactory } from './configFactory';
